+++ src/commands/GenerateContainer.ts
@@ -1,7 +1,7 @@
 import * as fs from "fs-extra";
 import * as path from "path";
 import { logger } from "../logger";
-import { targetPaths } from "../paths";
+import { pathExists, targetPaths } from "../paths";
 import { generateContainer } from "../templates/container";
 import { generateExport } from "../templates/export.t";
 import { capitalize } from "../templates/string.utils";
@@ -22,12 +22,15 @@ export class GenerateContainerCommand extends Command {
     }
 
     public action = (name: string) => {
+
+        if (!this.checkRequiredFiles()) {
+            return;
+        }
+
         const containerCode = generateContainer({ containerName: name });
         const exportCode = generateExport(`${capitalize(name)}Container`);
 
         logger.info(`Generating ${name} container ..`);
-
-        // TODO: Add check for duplicate component name
 
         const containerExportFolderPath = path.join(
             targetPaths.containers,
@@ -45,6 +48,11 @@ export class GenerateContainerCommand extends Command {
             `${capitalize(name)}Container`,
             `${capitalize(name)}Container.tsx`,
         );
+
+        if (pathExists(containerFilePath)) {
+            logger.error(`Container "${name}" already exists at path: ${containerFilePath}`);
+            return;
+        }
 
         // create & write file
         fs.outputFileSync(containerFilePath, containerCode);
